+++ apps/api/src/scraper/scrapeURL/engines/pdf/fire-pdf/poll.ts
@@ -1,0 +1,169 @@
+import type { Meta } from "../../..";
+import type { PDFProcessorResult } from "../types";
+import { fetch as undiciFetch } from "undici";
+import { AbortManagerThrownError } from "../../../lib/abortManager";
+import {
+  firePdfAsyncCompletedTotal,
+  firePdfAsyncPollCount,
+  type FallbackReason,
+} from "./metrics";
+import {
+  pollResponseSchema,
+  TERMINAL_STATUSES,
+  type PollResponse,
+} from "./schema";
+import { nextPollDelay, type Fallback } from "./utils";
+
+export type PollDeps = {
+  baseUrl: string;
+  scrapeId: string;
+  authHeader: Record<string, string>;
+  initialDelay: number;
+  pollingDeadline: number;
+  meta: Meta;
+  fetchImpl: typeof undiciFetch;
+  sleep: (ms: number, signal?: AbortSignal) => Promise<void>;
+  now: () => number;
+  fallback: Fallback;
+};
+
+export type PollOutcome =
+  | { kind: "done"; poll: PollResponse; pollCount: number }
+  | { kind: "fallback"; result: PDFProcessorResult };
+
+export async function pollUntilTerminal(
+  deps: PollDeps,
+): Promise<PollOutcome> {
+  const {
+    baseUrl,
+    scrapeId,
+    authHeader,
+    pollingDeadline,
+    meta,
+    fetchImpl,
+    sleep,
+    now,
+    fallback,
+  } = deps;
+  let pollCount = 0;
+  let lastDelay = deps.initialDelay;
+
+  while (true) {
+    if (now() > pollingDeadline) {
+      firePdfAsyncPollCount.observe(pollCount);
+      return {
+        kind: "fallback",
+        result: await fallback("polling_timeout", { pollCount }),
+      };
+    }
+
+    meta.abort.throwIfAborted();
+    await sleep(lastDelay, meta.abort.asSignal());
+    pollCount++;
+
+    let pollResp;
+    try {
+      pollResp = await fetchImpl(`${baseUrl}/jobs/${scrapeId}`, {
+        method: "GET",
+        headers: { ...authHeader },
+        signal: meta.abort.asSignal(),
+      });
+    } catch (error) {
+      if (error instanceof AbortManagerThrownError) throw error;
+      firePdfAsyncPollCount.observe(pollCount);
+      return {
+        kind: "fallback",
+        result: await fallback("network_error", {
+          error: String(error),
+          pollCount,
+        }),
+      };
+    }
+
+    const pollStatus = pollResp.status;
+    const pollBody = await pollResp.json().catch(() => ({}));
+
+    if (pollStatus === 404) {
+      firePdfAsyncPollCount.observe(pollCount);
+      throw new Error(
+        "fire-pdf async GET /jobs/:id 404: scrape_id missing after successful submit",
+      );
+    }
+
+    if (pollStatus === 410) {
+      firePdfAsyncPollCount.observe(pollCount);
+      const parsed = pollResponseSchema.safeParse(pollBody);
+      const status = parsed.success ? parsed.data.status : "expired";
+      firePdfAsyncCompletedTotal.labels(status).inc();
+      return {
+        kind: "fallback",
+        result: await fallback(
+          status === "cancelled" ? "terminal_cancelled" : "terminal_expired",
+          { status, pollCount, body: pollBody },
+        ),
+      };
+    }
+
+    if (pollStatus === 502) {
+      firePdfAsyncPollCount.observe(pollCount);
+      firePdfAsyncCompletedTotal.labels("failed").inc();
+      return {
+        kind: "fallback",
+        result: await fallback("terminal_failed", {
+          pollCount,
+          body: pollBody,
+        }),
+      };
+    }
+
+    if (pollStatus !== 200 && pollStatus !== 202) {
+      firePdfAsyncPollCount.observe(pollCount);
+      return {
+        kind: "fallback",
+        result: await fallback("http_5xx", {
+          status: pollStatus,
+          body: pollBody,
+          pollCount,
+        }),
+      };
+    }
+
+    const parsed = pollResponseSchema.safeParse(pollBody);
+    if (!parsed.success) {
+      firePdfAsyncPollCount.observe(pollCount);
+      return {
+        kind: "fallback",
+        result: await fallback("http_5xx", {
+          error: String(parsed.error),
+          body: pollBody,
+          pollCount,
+        }),
+      };
+    }
+
+    if (TERMINAL_STATUSES.has(parsed.data.status)) {
+      firePdfAsyncPollCount.observe(pollCount);
+      firePdfAsyncCompletedTotal.labels(parsed.data.status).inc();
+      if (parsed.data.status !== "done") {
+        const reason: FallbackReason =
+          parsed.data.status === "failed"
+            ? "terminal_failed"
+            : parsed.data.status === "expired"
+              ? "terminal_expired"
+              : "terminal_cancelled";
+        return {
+          kind: "fallback",
+          result: await fallback(reason, {
+            status: parsed.data.status,
+            errorClass: parsed.data.error_class,
+            errorMessage: parsed.data.error_message,
+            pollCount,
+          }),
+        };
+      }
+      return { kind: "done", poll: parsed.data, pollCount };
+    }
+
+    lastDelay = nextPollDelay(lastDelay, parsed.data.retry_after_ms);
+  }
+}
